EBookStore.Angular/src/app/core/progress-bar/progress-bar.component.ts
import { Component, OnInit, OnDestroy } from '@angular/core';
import { Subscription } from 'rxjs';

import { LoaderService, LoaderState } from '../../services/loader/loader.service';

@Component({
  selector: 'bs-progress-bar',
  templateUrl: './progress-bar.component.html',
  styleUrls: ['./progress-bar.component.scss']
})
export class BsProgressBarComponent implements OnInit, OnDestroy {
  show = false;
  private subscription: Subscription;

  constructor(private loaderService: LoaderService) { }

  ngOnInit() {
    this.subscription = this.loaderService.loaderState
    .subscribe((state: LoaderState) => {
      setTimeout(() => {
        this.show = state.show;
      }, 50);
    });
  }

  ngOnDestroy() {
    this.subscription.unsubscribe();
  }
}
